Add files via upload
this was stm32l433rct6p board with ec200u 4g module having the gps and gnss antenna. the code inside the file was taking the location details like lattitude and longitude after that sending sms to given mobile number in a certain time period mentioned in code for example 8 hours or any thershold time period.